refactor: NavigationBar.vue refactor of profile employee linking
